added Milestone 4

+++ Milestone1/routes/web.php
@@ -41,29 +41,40 @@ Route::get('/educationHistory', function(){
     return view('educationHistory');
 });
 
-Route::get('/loginRegister', 'loginRegisterController@show');
-Route::post('/login', 'loginRegisterController@login');
-Route::post('/register', 'loginRegisterController@register');
+Route::get('/groups', 'groupController@index');
+Route::get('/group/create', 'groupController@create')->middleware('auth');
+Route::post('/group/create', 'groupController@store')->middleware('auth');
+Route::get('/group/{group_id}/view', 'groupController@show');
+Route::get('/group/{group_id}/edit', 'groupController@edit')->middleware('auth');
+Route::patch('/group/{group_id}/edit', 'groupController@change')->middleware('auth');
+Route::delete('/group/{group_id}/delete', 'groupController@delete')->middleware('auth');
+Route::get('/group/{group_id}/leave', 'groupController@leave')->middleware('auth');
+Route::get('/group/{group_id}/join', 'groupController@join')->middleware('auth');
 
-Route::get('/profile/{userID}/create', 'profileController@create');
-Route::post('/profile/{userID}/create', 'profileController@addDemographics');
-Route::get('/profile/{userID}/workHistory', 'profileController@showWorkHistory');
-Route::post('/workHistory/{userID}/create', 'profileController@addWorkHistory');
-Route::get('/profile/{userID}/educationHistory', 'profileController@showEducationHistory');
-Route::post('/profile/{userID}/educationHistory', 'profileController@addEducationHistory');
+Auth::routes();
+Route::get('/logout', 'Auth\loginController@logout');
+
+
+Route::get('/profile/{userID}/create', 'profileController@create')->middleware('auth');
+Route::post('/profile/{userID}/create', 'profileController@addDemographics')->middleware('auth');
+Route::get('/profile/{userID}/workHistory', 'profileController@showWorkHistory')->middleware('auth');
+Route::post('/workHistory/{userID}/create', 'profileController@addWorkHistory')->middleware('auth');
+Route::get('/profile/{userID}/educationHistories', 'profileController@showEducationHistory')->middleware('auth');
+Route::post('/profile/{userID}/educationHistories', 'profileController@addEducationHistory')->middleware('auth');
 Route::get('/profile/{userID}/view', 'profileController@show');
 
+Route::get('/admin/users', 'adminController@showUsers')->middleware('can:access');
+Route::patch('/users/suspend', 'adminController@suspendUser')->middleware('can:access');
+Route::delete('/users/delete', 'adminController@deleteUser')->middleware('can:access');
+Route::post('/users/reactivate', 'adminController@reactivateUser')->middleware('can:access');
+Route::patch('/users/admin', 'adminController@toAdmin')->middleware('can:access');
+Route::post('/users/admin', 'adminController@fromAdmin')->middleware('can:access');
 
-Route::get('/admin/users', 'adminController@showUsers');
-Route::patch('/users/suspend', 'adminController@suspendUser');
-Route::delete('/users/delete', 'adminController@deleteUser');
-Route::post('/users/reactivate', 'adminController@reactivateUser');
-Route::patch('/users/admin', 'adminController@toAdmin');
-Route::post('/users/admin', 'adminController@fromAdmin');
+Route::get('/jobs', 'jobsController@show')->middleware('auth');
+Route::get('/admin/jobs/add', 'jobsController@showAdd')->middleware('can:access');
+Route::post('/admin/jobs/add', 'jobsController@addJob')->middleware('can:access');
+Route::patch('/admin/jobs/{jobID}/edit', 'jobsController@editJob')->middleware('can:access');
+Route::get('/jobs/{jobID}/view', 'jobsController@showJob')->middleware('auth');
+Route::delete('/jobs/{jobID}/delete', 'jobsController@deleteJob')->middleware('can:access');
 
-Route::get('/admin/jobs', 'jobsController@show');
-Route::get('/admin/jobs/add', 'jobsController@showAdd');
-Route::post('/admin/jobs/add', 'jobsController@addJob');
-Route::patch('/admin/jobs/{jobID}/edit', 'jobsController@editJob');
-Route::get('/jobs/{jobID}/view', 'jobsController@showJob');
-Route::delete('/jobs/{jobID}/delete', 'jobsController@deleteJob');
+Route::get('/home', 'HomeController@index')->name('home');
